refactor: complete status application extraction
